+++ common/src/main/java/com/slymask3/instantblocks/builder/BlockPosHolder.java
@@ -7,12 +7,17 @@ import java.util.List;
 
 public class BlockPosHolder {
     private final List<BlockPos> posList;
-    private final CheckDirections checkDirections;
     private final CheckBlock checkBlock;
+    private final boolean north, east, south, west, above, below;
 
-    public BlockPosHolder(BlockPos origin, CheckDirections checkDirections, CheckBlock checkBlock) {
+    public BlockPosHolder(BlockPos origin, boolean north, boolean east, boolean south, boolean west, boolean above, boolean below, CheckBlock checkBlock) {
         this.posList = new ArrayList<>();
-        this.checkDirections = checkDirections;
+        this.north = north;
+        this.east = east;
+        this.south = south;
+        this.west = west;
+        this.above = above;
+        this.below = below;
         this.checkBlock = checkBlock;
         this.checkDirections(origin);
         this.checkBlock(origin);
@@ -39,15 +44,16 @@ public class BlockPosHolder {
     }
 
     public void checkDirections(BlockPos pos) {
-        this.checkDirections.call(pos, this);
+        if(this.north)this.checkBlock(pos.north(1));
+        if(this.east)this.checkBlock(pos.east(1));
+        if(this.south)this.checkBlock(pos.south(1));
+        if(this.west)this.checkBlock(pos.west(1));
+        if(this.above)this.checkBlock(pos.above(1));
+        if(this.below)this.checkBlock(pos.below(1));
     }
 
     public void checkBlock(BlockPos pos) {
         this.checkBlock.call(pos, this);
-    }
-
-    public interface CheckDirections {
-        void call(BlockPos pos, BlockPosHolder holder);
     }
 
     public interface CheckBlock {