+++ common/src/main/java/com/slymask3/instantblocks/builder/BuildSound.java
@@ -1,5 +1,6 @@
 package com.slymask3.instantblocks.builder;
 
+import com.slymask3.instantblocks.util.ClientHelper;
 import net.minecraft.core.BlockPos;
 import net.minecraft.sounds.SoundEvent;
 
@@ -7,12 +8,14 @@ public class BuildSound {
     private final BlockPos pos;
     private final SoundEvent placeSound, breakSound;
     private final float volume;
+    private final ClientHelper.Particles particles;
 
-    public BuildSound(BlockPos pos, SoundEvent placeSound, SoundEvent breakSound, float volume) {
+    public BuildSound(BlockPos pos, SoundEvent placeSound, SoundEvent breakSound, float volume, ClientHelper.Particles particles) {
         this.pos = pos;
         this.placeSound = placeSound;
         this.breakSound = breakSound;
         this.volume = volume;
+        this.particles = particles;
     }
 
     public BlockPos getBlockPos() {
@@ -37,5 +40,9 @@ public class BuildSound {
 
     public float getVolume() {
         return this.volume;
+    }
+
+    public ClientHelper.Particles getParticles() {
+        return this.particles;
     }
 }